+++ CHANGELOG.md
@@ -5,11 +5,33 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.24.2] - 2026-06-30
+
+### Fixed
+
+- Make "Associated item type" mandatory when creating a container.
+
+## [1.24.1] - 2026-06-24
+
+### Fixed
+
+- Fix the name of dropdown fields in logs when updating an item
+- Fix plugin configuration deletion during uninstallation
+- Fix nested array stored in DB for readonly dropdown
+- Fix migration abort when a GenericObject container name produces a table name exceeding MySQL's 64-character limit after conversion to GlpiCustomAsset.
+- Fix empty dropdown value (-1) on form submission
+- Fix text area fields size and alignment
+- Optimize container loading when there are a large number of entities
+- Adding a verification in refreshContainer function for obj value which can be an empty string instead of an array
+- Fix refreshContainer crash when a field is serialized both as a scalar and as an array
+- Centralized label preparation and system name generation.
+
 ## [1.24.0] - 2026-04-16
 
 ### Added
 
 - Support for new value operators for “Field” question type linked to a field that is a dropdown in forms
+- Support for searching future dates and times in date/datetime fields
 
 ### Fixed